requirements.txt
# Keep your existing working dependencies (don't change these)
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
async-timeout==5.0.1
attrs==25.3.0
cachetools==5.5.2
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.1.8
colorama==0.4.6
distro==1.9.0
Django==5.2.2
eval_type_backport==0.2.2
exceptiongroup==1.3.0
filelock==3.18.0
frozenlist==1.7.0
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.173.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
googleapis-common-protos==1.70.0
grpcio==1.73.0
grpcio-status==1.71.0
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
idna==3.10
jiter==0.10.0
markdown-it-py==3.0.0
mdurl==0.1.2
multidict==6.5.0
numpy==2.2.6
packaging==25.0
pillow==11.2.1
propcache==0.3.2
proto-plus==1.26.1
protobuf==5.29.5
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.1
pyparsing==3.2.3
python-dotenv==1.1.0
requests==2.32.4
rich==14.0.0
rsa==4.9.1
shellingham==1.5.4
sniffio==1.3.1
sqlparse==0.5.3
tabulate==0.9.0
together==1.5.16
tqdm==4.67.1
typer==0.15.4
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
yarl==1.20.1

# Add only essential ELK dependencies (compatible versions)
elasticsearch>=7.0.0,<9.0.0
redis>=4.0.0,<6.0.0

# AI Services (compatible versions)
openai>=1.0.0,<2.0.0

# Basic utilities
python-dateutil>=2.8.0
requests-oauthlib>=1.3.0